chore(firefox): move sources from //lib/firefox to //lib
