+++ experimental/puppeteer-firefox/lib/Browser.js
