+++ experimental/puppeteer-firefox/lib/BrowserFetcher.js
