+++ experimental/puppeteer-firefox/lib/Connection.js
