+++ experimental/puppeteer-firefox/lib/Dialog.js
