+++ experimental/puppeteer-firefox/lib/FirefoxTransport.js
