+++ experimental/puppeteer-firefox/lib/Input.js
