+++ experimental/puppeteer-firefox/lib/USKeyboardLayout.js
