+++ experimental/puppeteer-firefox/lib/externs.d.ts
